+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -2,8 +2,10 @@ CHANGELOG
 =========
 
 4.4.0
+-----
 
-* Deprecated the usage of "query_string" without a "search_dn" and a "search_password" config key in Ldap factories.
+ * Added new `argon2id` encoder, undeprecated the `bcrypt` and `argon2i` ones (using `auto` is still recommended by default.)
+ * Deprecated the usage of "query_string" without a "search_dn" and a "search_password" config key in Ldap factories.
 
 4.3.0
 -----
@@ -14,7 +16,6 @@ CHANGELOG
    option is deprecated and will be disabled in Symfony 5.0. This affects to cookies
    with dashes in their names. For example, starting from Symfony 5.0, the `my-cookie`
    name will delete `my-cookie` (with a dash) instead of `my_cookie` (with an underscore).
- * Deprecated configuring encoders using `argon2i` as algorithm, use `auto` instead
 
 4.2.0
 -----